Introduce gammapy/irf/psf/parametric.py
