+++ gammapy/irf/psf/__init__.py
@@ -1,5 +1,4 @@
-from .gauss import *
 from .kernel import *
-from .king import *
+from .parametric import *
 from .psf_map import *
 from .table import *
