+++ gammapy/irf/psf/core.py
@@ -1,10 +1,8 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-import abc
 import numpy as np
 from astropy import units as u
 from gammapy.maps import MapAxes, MapAxis
 from gammapy.utils.array import array_stats_str
-from gammapy.utils.interpolation import ScaledRegularGridInterpolator
 from ..core import IRF
 
 
@@ -262,184 +260,3 @@ class PSF(IRF):
         self.plot_containment(fraction=0.95, ax=axes[1])
         self.plot_containment_vs_energy(ax=axes[2])
         plt.tight_layout()
-
-
-class ParametricPSF(PSF):
-    """Parametric PSF base class
-
-    Parameters
-    -----------
-    axes : list of `MapAxis` or `MapAxes`
-        Axes
-    data : dict of `~numpy.ndarray`, or `~numpy.recarray`
-        Data
-    unit : dict of str or `~astropy.units.Unit`
-        Unit
-    meta : dict
-        Meta data
-    """
-    @property
-    @abc.abstractmethod
-    def required_parameters(self):
-        pass
-
-    @abc.abstractmethod
-    def evaluate(self, rad):
-        pass
-
-    @property
-    def quantity(self):
-        """Quantity"""
-        quantity = {}
-
-        for name in self.required_parameters:
-            quantity[name] = self.data[name] * self.unit[name]
-
-        return quantity
-
-    @property
-    def unit(self):
-        """Map unit (`~astropy.units.Unit`)"""
-        return self._unit
-
-    @unit.setter
-    def unit(self, values):
-        self._unit = {key: u.Unit(val) for key, val in values.items()}
-
-    @property
-    def _interpolators(self):
-        interps = {}
-
-        for name in self.required_parameters:
-            points = [a.center for a in self.axes]
-            points_scale = tuple([a.interp for a in self.axes])
-            interps[name] = ScaledRegularGridInterpolator(
-                points, values=self.quantity[name], points_scale=points_scale
-            )
-
-        return interps
-
-    def evaluate_parameters(self, energy_true, offset):
-        """Evaluate analytic PSF parameters at a given energy and offset.
-
-        Uses nearest-neighbor interpolation.
-
-        Parameters
-        ----------
-        energy_true : `~astropy.units.Quantity`
-            energy value
-        offset : `~astropy.coordinates.Angle`
-            Offset in the field of view
-
-        Returns
-        -------
-        values : `~astropy.units.Quantity`
-            Interpolated value
-        """
-        pars = {}
-        for name in self.required_parameters:
-            value = self._interpolators[name]((energy_true, offset))
-            pars[name] = value
-
-        return pars
-
-    def to_table(self, format="gadf-dl3"):
-        """Convert PSF table data to table.
-
-        Parameters
-        ----------
-        format : {"gadf-dl3"}
-            Format specification
-
-
-        Returns
-        -------
-        hdu_list : `~astropy.io.fits.HDUList`
-            PSF in HDU list format.
-        """
-        table = self.axes.to_table(format="gadf-dl3")
-
-        for name in self.required_parameters:
-            table[name.upper()] = self.data[name].T[np.newaxis]
-            table[name.upper()].unit = self.unit[name]
-
-        # Create hdu and hdu list
-        return table
-
-    @classmethod
-    def from_table(cls, table, format="gadf-dl3"):
-        """Create parametric psf from `~astropy.table.Table`.
-
-        Parameters
-        ----------
-        table : `~astropy.table.Table`
-            Table  info.
-
-        Returns
-        -------
-        psf : `~ParametricPSF`
-            PSF class
-        """
-        axes = MapAxes.from_table(table, format=format)[cls.required_axes]
-
-        dtype = {
-            "names": cls.required_parameters,
-            "formats": len(cls.required_parameters) * (np.float32,)
-        }
-
-        data = np.empty(axes.shape, dtype=dtype)
-        unit = {}
-
-        for name in cls.required_parameters:
-            column = table[name.upper()]
-            values = column.data[0].transpose()
-
-            # TODO: this fixes some files where sigma is written as zero
-            if "SIGMA" in name:
-                values[values == 0] = 1.
-
-            data[name] = values.reshape(axes.shape)
-            unit[name] = column.unit or ""
-
-        return cls(
-            axes=axes,
-            data=data,
-            meta=table.meta.copy(),
-            unit=unit
-        )
-
-    def to_psf3d(self, rad=None):
-        """Create a PSF3D from a parametric PSF.
-
-        It will be defined on the same energy and offset values than the input psf.
-
-        Parameters
-        ----------
-        rad : `~astropy.units.Quantity`
-            Rad values
-
-        Returns
-        -------
-        psf3d : `~gammapy.irf.PSF3D`
-            PSF3D.
-        """
-        from gammapy.irf import PSF3D
-        from gammapy.datasets.map import RAD_AXIS_DEFAULT
-
-        offset_axis = self.axes["offset"]
-        energy_axis_true = self.axes["energy_true"]
-
-        if rad is None:
-            rad_axis = RAD_AXIS_DEFAULT.center
-        else:
-            rad_axis = MapAxis.from_edges(rad, name="rad")
-
-        axes = MapAxes([energy_axis_true, offset_axis, rad_axis])
-        data = self.evaluate(**axes.get_coord())
-
-        return PSF3D(
-            axes=axes,
-            data=data.value,
-            unit=data.unit,
-            meta=self.meta.copy()
-        )
